move px.FBO into jarsGl as `framebuffer`
the jarsGl.framebuffer object is just that, an object, not a
constructor. therefore it is now instantiated via a call to
Object.create(jarsGl.framebuffer). the only function call previously in
the constructor was to construct the pxBB object, which was moved to
jarsGl.framebuffer.allocate(). that object is next for rewriting.

additionally, all its methods now require the webgl rendering context as
a parameter, rather than having that as a property on the object.